refactor: move presets into their own class
This commit moves presets into their own class. A preset can be loaded
from any valid preset file (see data/presets for examples) or the
configuration file (pass the config=True keyword parameter) with the
static from_file method.

The from_file method also allows to pass the boolean system keyword to
mark it as system preset.

In addition quiet allot of globals have been eliminated.
GetSettings now returns the preset loaded from the settings.
ApplySettings now expects a preset as parameter to write to the config
file.

The global preset moved to Equalizer.preset which keeps track of the
current preset and passes it to ApplySettings when needed or is set to
the GetSettings returned preset when settings are read.

Preset specific globals are not needed anymore since they are part of
the preset class now and mapped as following:
 * ladspa_filename -> preset.plugin
 * ladspa_name -> preset.plugin_name
 * ladspa_label -> preset.plugin_label
 * num_ladspa_controls -> len(preset.Bands)
 * ladspa_controls[idx] -> preset.Bands[idx].control
 * ladspa_inputs[idx] -> preset.Bands[idx].frequency

The rawpresets list has moved to Equalizer.presets and contains already
read presets. Note user presets that have the same name as system
presets currently act as overrides.

The global clearpreset has moved to Equalizer.update_preset. It acts as
guard for Equalizer.on_scale so we only set the presets name to '' when
the user changes a scale value.

The global presetmatch has been removed in favor of just checking if
Equalizer.presets contains a certain preset name.

The ComboBoxText has been replaced with a plain ComboBox with the
ListStore Equalizer.presetsstore set as model. The store has the columns
preset.name and preset which allows to load a preset directly associated
to an entry.